refactor(project): restructure directory - move all module to the folder modoules
