+++ src/modules/plants/dto/create-plant.dto.ts
