+++ src/modules/plants/plants.service.ts
